# docs/swagger.yaml
info:
  contact: {}
paths:
  /api/v1/tag/{id}:
    delete:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":10002,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: Delete Tag
    put:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: modifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":10001,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: Edit Tag
  /api/v1/tags:
    get:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: Get multiple article tags
    post:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: integer
      - description: createdBy
        in: query
        name: created_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":10001,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: Add Tag
swagger: "2.0"